HGR_API/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
imutils==0.5.4
matplotlib==3.3.4
numpy==1.18.4
opencv-python==3.4.3.18
scikit-image==0.15.0
scikit-learn==0.22.2
tensorflow==2.3.0
tensorflow-serving-api==2.3.0
gunicorn==20.0.4
requests==2.23.0
easydict==1.9
